add: tiled tensor mul

+++ src/cuda/storage.cu
@@ -11,18 +11,16 @@
 
 Storage::Storage(thrust::host_vector<std::size_t> shape, float value = 0)
     : shape(shape) {
-  std::size_t size =
-      thrust::reduce(shape.begin(), shape.end(), (std::size_t)1,
-                     thrust::multiplies<std::size_t>());
+  std::size_t size = thrust::reduce(shape.begin(), shape.end(), (std::size_t)1,
+                                    thrust::multiplies<std::size_t>());
   this->data.resize(size, value);
 }
 
 Storage::Storage(thrust::host_vector<std::size_t> shape,
                  thrust::device_vector<float> &&data)
     : shape(shape) {
-  std::size_t size =
-      thrust::reduce(shape.begin(), shape.end(), (std::size_t)1,
-                     thrust::multiplies<std::size_t>());
+  std::size_t size = thrust::reduce(shape.begin(), shape.end(), (std::size_t)1,
+                                    thrust::multiplies<std::size_t>());
   if (data.size() != size) {
     throw "data size != shape";
   }
@@ -33,9 +31,8 @@ Storage::Storage(thrust::host_vector<std::size_t> shape,
 void Storage::xavier(size_t in_size, size_t out_size) {
   float *a_ptr = thrust::raw_pointer_cast(this->data.data());
   std::size_t size = this->data.size();
-  std::size_t block_size = ceil(std::static_cast<double>(size) / BLOCK_SIZE);
-  float scale = std::sqrt((double)6) /
-                std::sqrt(std::static_cast<double>(in_size) + out_size);
+  std::size_t block_size = ceil((float)(size) / BLOCK_SIZE);
+  float scale = std::sqrt((double)6) / std::sqrt((float)(in_size) + out_size);
   storage_xavier<<<block_size, BLOCK_SIZE>>>(a_ptr, size, scale);
 }
